refactor: TiptapEditor 컴포넌트 파일 분리
단일 책임 원칙에 따라 5개 관심사를 별도 파일로 분리:
- AlignableImage/AlignableYoutube → extensions/
- MenuBar, MediaBubbleMenu → 독립 컴포넌트
- moveNode → utils/

#7
