+++ src/components/admin/TiptapEditor.tsx
@@ -1,415 +1,19 @@
 'use client';
 
-import Image from '@tiptap/extension-image';
 import Placeholder from '@tiptap/extension-placeholder';
-import { Fragment } from '@tiptap/pm/model';
-import { NodeSelection } from '@tiptap/pm/state';
-import Youtube from '@tiptap/extension-youtube';
-import { type Editor, EditorContent, useEditor } from '@tiptap/react';
-import { BubbleMenu } from '@tiptap/react/menus';
+import { EditorContent, useEditor } from '@tiptap/react';
 import StarterKit from '@tiptap/starter-kit';
-import { useCallback, useState } from 'react';
 
-import { useToast } from '@/components/ui/Toast';
-import { ALIGN, NODE_TYPE } from '@/constants/editor';
-import { uploadImage } from '@/lib/image';
-import { createClient } from '@/lib/supabase/client';
+import AlignableImage from './extensions/AlignableImage';
+import AlignableYoutube from './extensions/AlignableYoutube';
+import MediaBubbleMenu from './MediaBubbleMenu';
+import MenuBar from './MenuBar';
 
-const MAX_IMAGE_WIDTH = 1200;
 const MIN_RESIZE_SIZE = 100;
 
 interface TiptapEditorProps {
   content: string;
   onChange: (html: string) => void;
-}
-
-const YOUTUBE_URL_PATTERN =
-  /^https?:\/\/(www\.)?(youtube\.com\/watch\?v=|youtu\.be\/|youtube\.com\/embed\/)/;
-
-const AlignableImage = Image.extend({
-  addAttributes() {
-    return {
-      ...this.parent?.(),
-      'data-align': {
-        default: null,
-        parseHTML: (el: HTMLElement) => el.getAttribute('data-align'),
-        renderHTML: (attrs: Record<string, unknown>) => {
-          if (!attrs['data-align']) return {};
-          return { 'data-align': attrs['data-align'] as string };
-        },
-      },
-    };
-  },
-  addNodeView() {
-    const parentRenderer = this.parent?.();
-    if (!parentRenderer) return null;
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    return (props: any) => {
-      const view = parentRenderer(props);
-
-      const applyAlign = (align: string | null) => {
-        if (view.dom instanceof HTMLElement) {
-          if (align) view.dom.setAttribute('data-align', align);
-          else view.dom.removeAttribute('data-align');
-        }
-      };
-
-      applyAlign(props.node.attrs['data-align'] || null);
-
-      const origUpdate = view.update;
-      if (origUpdate) {
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-        view.update = (node: any, decorations: any, innerDecorations: any) => {
-          const result = origUpdate.call(
-            view,
-            node,
-            decorations,
-            innerDecorations,
-          );
-          if (result !== false) {
-            applyAlign(node?.attrs?.['data-align'] || null);
-          }
-          return result;
-        };
-      }
-
-      return view;
-    };
-  },
-});
-
-const AlignableYoutube = Youtube.extend({
-  addAttributes() {
-    return {
-      ...this.parent?.(),
-      'data-align': {
-        default: null,
-        parseHTML: (el: HTMLElement) => el.getAttribute('data-align'),
-        renderHTML: (attrs: Record<string, unknown>) => {
-          if (!attrs['data-align']) return {};
-          return { 'data-align': attrs['data-align'] as string };
-        },
-      },
-    };
-  },
-});
-
-function MenuBar({ editor }: { editor: ReturnType<typeof useEditor> }) {
-  const { toast } = useToast();
-  const [showYoutubeInput, setShowYoutubeInput] = useState(false);
-  const [youtubeUrl, setYoutubeUrl] = useState('');
-
-  if (!editor) return null;
-
-  const btnClass = (active: boolean) =>
-    `rounded px-2 py-1 text-sm transition-colors ${
-      active
-        ? 'bg-accent text-white'
-        : 'text-muted hover:text-foreground hover:bg-foreground/5'
-    }`;
-
-  const handleImageUpload = async () => {
-    const input = document.createElement('input');
-    input.type = 'file';
-    input.accept = 'image/*';
-    input.onchange = async () => {
-      const file = input.files?.[0];
-      if (!file) return;
-      try {
-        const supabase = createClient();
-        const url = await uploadImage(supabase, file, MAX_IMAGE_WIDTH, 'posts');
-        editor
-          .chain()
-          .focus()
-          .setImage({ src: url })
-          .createParagraphNear()
-          .run();
-      } catch {
-        toast('이미지 업로드에 실패했습니다.');
-      }
-    };
-    input.click();
-  };
-
-  const handleYoutubeSubmit = () => {
-    const trimmed = youtubeUrl.trim();
-    if (!trimmed) return;
-
-    if (!YOUTUBE_URL_PATTERN.test(trimmed)) {
-      toast('유효한 YouTube URL을 입력해주세요.');
-      return;
-    }
-
-    editor
-      .chain()
-      .focus()
-      .setYoutubeVideo({ src: trimmed })
-      .createParagraphNear()
-      .run();
-    setYoutubeUrl('');
-    setShowYoutubeInput(false);
-  };
-
-  return (
-    <div className="border-b border-border p-2">
-      <div className="flex flex-wrap gap-1">
-        <button
-          type="button"
-          onClick={() =>
-            editor.chain().focus().toggleHeading({ level: 2 }).run()
-          }
-          className={btnClass(editor.isActive('heading', { level: 2 }))}
-        >
-          제목2
-        </button>
-        <button
-          type="button"
-          onClick={() =>
-            editor.chain().focus().toggleHeading({ level: 3 }).run()
-          }
-          className={btnClass(editor.isActive('heading', { level: 3 }))}
-        >
-          제목3
-        </button>
-        <button
-          type="button"
-          onClick={() => editor.chain().focus().toggleBold().run()}
-          className={btnClass(editor.isActive('bold'))}
-        >
-          굵게
-        </button>
-        <button
-          type="button"
-          onClick={() => editor.chain().focus().toggleItalic().run()}
-          className={btnClass(editor.isActive('italic'))}
-        >
-          기울임
-        </button>
-        <button
-          type="button"
-          onClick={() => editor.chain().focus().toggleBulletList().run()}
-          className={btnClass(editor.isActive('bulletList'))}
-          aria-label="순서 없는 목록"
-        >
-          ● 목록
-        </button>
-        <button
-          type="button"
-          onClick={() => editor.chain().focus().toggleOrderedList().run()}
-          className={btnClass(editor.isActive('orderedList'))}
-          aria-label="순서 있는 목록"
-        >
-          1. 목록
-        </button>
-        <button
-          type="button"
-          onClick={() => editor.chain().focus().toggleBlockquote().run()}
-          className={btnClass(editor.isActive('blockquote'))}
-        >
-          인용
-        </button>
-        <button
-          type="button"
-          onClick={handleImageUpload}
-          className={btnClass(false)}
-        >
-          이미지
-        </button>
-        <button
-          type="button"
-          onClick={() => setShowYoutubeInput(!showYoutubeInput)}
-          className={btnClass(showYoutubeInput)}
-        >
-          유튜브
-        </button>
-      </div>
-      {showYoutubeInput && (
-        <div className="mt-2 flex gap-2">
-          <input
-            type="url"
-            value={youtubeUrl}
-            onChange={(e) => setYoutubeUrl(e.target.value)}
-            onKeyDown={(e) =>
-              e.key === 'Enter' && (e.preventDefault(), handleYoutubeSubmit())
-            }
-            placeholder="https://www.youtube.com/watch?v=..."
-            className="flex-1 rounded border border-border bg-background px-2 py-1 text-sm text-foreground focus:outline-none focus:ring-1 focus:ring-accent"
-          />
-          <button
-            type="button"
-            onClick={handleYoutubeSubmit}
-            className="rounded bg-accent px-3 py-1 text-sm text-white hover:bg-accent/80 transition-colors"
-          >
-            삽입
-          </button>
-          <button
-            type="button"
-            onClick={() => {
-              setShowYoutubeInput(false);
-              setYoutubeUrl('');
-            }}
-            className="rounded px-2 py-1 text-sm text-muted hover:text-foreground transition-colors"
-          >
-            취소
-          </button>
-        </div>
-      )}
-    </div>
-  );
-}
-
-function moveNode(editor: Editor, direction: 'up' | 'down'): boolean {
-  const { state } = editor;
-  const { selection } = state;
-
-  if (!(selection instanceof NodeSelection)) return false;
-
-  const node = selection.node;
-  const pos = selection.from;
-  const $pos = state.doc.resolve(pos);
-  const index = $pos.index();
-  const parent = $pos.parent;
-
-  if (direction === 'up' && index === 0) return false;
-  if (direction === 'down' && index >= parent.childCount - 1) return false;
-
-  const tr = state.tr;
-
-  if (direction === 'up') {
-    const prevNode = parent.child(index - 1);
-    const prevStart = pos - prevNode.nodeSize;
-    tr.replaceWith(
-      prevStart,
-      pos + node.nodeSize,
-      Fragment.fromArray([node, prevNode]),
-    );
-    tr.setSelection(NodeSelection.create(tr.doc, prevStart));
-  } else {
-    const nextNode = parent.child(index + 1);
-    const nextEnd = pos + node.nodeSize + nextNode.nodeSize;
-    tr.replaceWith(pos, nextEnd, Fragment.fromArray([nextNode, node]));
-    tr.setSelection(NodeSelection.create(tr.doc, pos + nextNode.nodeSize));
-  }
-
-  editor.view.dispatch(tr);
-  return true;
-}
-
-function MediaBubbleMenu({ editor }: { editor: Editor }) {
-  const btnClass = (active: boolean) =>
-    `rounded px-2 py-1 text-sm transition-colors ${
-      active
-        ? 'bg-accent/20 text-accent'
-        : 'text-muted hover:text-foreground hover:bg-foreground/5'
-    }`;
-
-  const getActiveNodeType = useCallback((): string | null => {
-    if (editor.isActive(NODE_TYPE.IMAGE)) return NODE_TYPE.IMAGE;
-    if (editor.isActive(NODE_TYPE.YOUTUBE)) return NODE_TYPE.YOUTUBE;
-    return null;
-  }, [editor]);
-
-  const getCurrentAlign = useCallback((): string | null => {
-    const nodeType = getActiveNodeType();
-    if (!nodeType) return null;
-    return editor.getAttributes(nodeType)['data-align'] || null;
-  }, [editor, getActiveNodeType]);
-
-  const handleAlign = useCallback(
-    (align: string | null) => {
-      const nodeType = getActiveNodeType();
-      if (!nodeType) return;
-      editor.commands.updateAttributes(nodeType, { 'data-align': align });
-    },
-    [editor, getActiveNodeType],
-  );
-
-  const handleMoveUp = useCallback(() => {
-    moveNode(editor, 'up');
-  }, [editor]);
-
-  const handleMoveDown = useCallback(() => {
-    moveNode(editor, 'down');
-  }, [editor]);
-
-  const handleDelete = useCallback(() => {
-    editor.commands.deleteSelection();
-  }, [editor]);
-
-  const currentAlign = getCurrentAlign();
-
-  return (
-    <BubbleMenu
-      editor={editor}
-      shouldShow={({ editor: e }) =>
-        e.isActive(NODE_TYPE.IMAGE) || e.isActive(NODE_TYPE.YOUTUBE)
-      }
-    >
-      <div className="flex items-center gap-1 rounded-lg border border-border bg-card p-1 shadow-lg">
-        <div role="group" aria-label="정렬" className="flex items-center gap-1">
-          <button
-            type="button"
-            onClick={() => handleAlign(null)}
-            className={btnClass(!currentAlign)}
-            title="왼쪽 정렬"
-            aria-label="왼쪽 정렬"
-            aria-pressed={!currentAlign}
-          >
-            좌
-          </button>
-          <button
-            type="button"
-            onClick={() => handleAlign(ALIGN.CENTER)}
-            className={btnClass(currentAlign === ALIGN.CENTER)}
-            title="가운데 정렬"
-            aria-label="가운데 정렬"
-            aria-pressed={currentAlign === ALIGN.CENTER}
-          >
-            중
-          </button>
-          <button
-            type="button"
-            onClick={() => handleAlign(ALIGN.RIGHT)}
-            className={btnClass(currentAlign === ALIGN.RIGHT)}
-            title="오른쪽 정렬"
-            aria-label="오른쪽 정렬"
-            aria-pressed={currentAlign === ALIGN.RIGHT}
-          >
-            우
-          </button>
-        </div>
-        <div className="mx-0.5 h-5 w-px bg-border" />
-        <button
-          type="button"
-          onClick={handleMoveUp}
-          className={btnClass(false)}
-          title="위로 이동"
-          aria-label="위로 이동"
-        >
-          ↑
-        </button>
-        <button
-          type="button"
-          onClick={handleMoveDown}
-          className={btnClass(false)}
-          title="아래로 이동"
-          aria-label="아래로 이동"
-        >
-          ↓
-        </button>
-        <button
-          type="button"
-          onClick={handleDelete}
-          className="rounded px-2 py-1 text-sm text-red-400 hover:text-red-300 hover:bg-red-400/10 transition-colors"
-          title="삭제"
-          aria-label="삭제"
-        >
-          ✕
-        </button>
-      </div>
-    </BubbleMenu>
-  );
 }
 
 export default function TiptapEditor({ content, onChange }: TiptapEditorProps) {
